# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления внутренней видео документацией Vilib.",
        "title": "Vilib API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/accounts/{accountId}/roles": {
            "get": {
                "description": "Возвращает список всех ролей аккаунта",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account_roles"
                ],
                "summary": "Список ролей аккаунта",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.AccountRole"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт новую роль для аккаунта с указанными правами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account_roles"
                ],
                "summary": "Создание роли аккаунта",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса для создания роли",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAccountRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAccountRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            }
        },
...
        "/api/v1/accounts/{accountId}/user-groups/roles": {
            "get": {
                "description": "Возвращает список всех ролей групп для аккаунта",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group_roles"
                ],
                "summary": "Список ролей групп",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.GroupRole"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт новую роль для группы пользователей с указанными правами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group_roles"
                ],
                "summary": "Создание роли группы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса для создания роли группы",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateGroupRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateGroupRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            }
        },
...
        "/api/v1/accounts/{accountId}/user-groups/{groupId}/video/{videoId}": {
            "get": {
                "description": "Возвращает преподписанный URL для доступа к видео",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Получение видео",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID группы пользователей",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID видео",
                        "name": "videoId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Предпочитать оригинальное видео",
                        "name": "prefer_original",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetVideoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            }
        },
...
        "/api/v1/accounts/{accountId}/user-groups/{userGroupId}/video/{videoId}": {
            "put": {
                "description": "Переименовывает видео в группе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Переименование видео",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID группы",
                        "name": "userGroupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID видео",
                        "name": "videoId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RenameVideoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RenameVideoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет видео из группы",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Удаление видео",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID группы",
                        "name": "userGroupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID видео",
                        "name": "videoId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.PermissionMask": {
            "type": "integer",
            "format": "int64",
            "enum": [
                0
            ],
            "x-enum-varnames": [
                "DefaultPermissionMask"
            ]
        },
        "dto.AccountRole": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "permission": {
                    "$ref": "#/definitions/domain.PermissionMask"
                }
            }
        },
...
        "dto.CreateAccountRoleRequest": {
            "type": "object",
            "properties": {
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "permission": {
                    "$ref": "#/definitions/domain.PermissionMask"
                }
            }
        },
        "dto.CreateAccountRoleResponse": {
            "type": "object",
            "properties": {
                "account_role": {
                    "$ref": "#/definitions/dto.AccountRole"
                }
            }
        },
        "dto.CreateGroupRoleRequest": {
            "type": "object",
            "properties": {
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "permission_mask": {
                    "$ref": "#/definitions/domain.PermissionMask"
                }
            }
        },
        "dto.CreateGroupRoleResponse": {
            "type": "object",
            "properties": {
                "role": {
                    "$ref": "#/definitions/dto.GroupRole"
                }
            }
        },
...
        "dto.ErrorMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.GetVideoResponse": {
            "type": "object",
            "properties": {
                "presigned_url": {
                    "type": "string"
                }
            }
        },
...
        "dto.GroupRole": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "permission_mask": {
                    "$ref": "#/definitions/domain.PermissionMask"
                }
            }
        },
...
        "dto.RenameVideoRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
        "dto.RenameVideoResponse": {
            "type": "object",
            "properties": {
                "video": {
                    "$ref": "#/definitions/dto.Video"
                }
            }
        },
...
        "dto.Video": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}